refactor: Rename and tidy up `scoring.ts`
