+++ src/app/tournament/[uuid]/_components/PlayerList.tsx
@@ -6,7 +6,7 @@ import { useAttendanceMutations } from "@/hooks/attendance/useAttendanceMutation
 import { useLogs } from "@/hooks/logs/useLogs";
 import { usePlayers } from "@/hooks/players/usePlayers";
 import { useTables } from "@/hooks/tables/useTables";
-import { rankPlayers, scoreToColor } from "@/lib/scores";
+import { rankPlayers, scoreToColor } from "@/lib/scoring";
 import type { Player } from "@/lib/types";
 import { useSessionContext } from "@/providers/SessionProvider";
